+++ src/hedux/component/HeduxContext.ts
@@ -1,5 +1,8 @@
 import { createContext } from 'react';
-import { Hedux } from '../core/types';
+import { HeduxStore } from '../core/types';
+import { Subscription } from '../utils/subscription';
+
+export type HeduxContextValue = HeduxStore & Subscription;
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const HeduxContext = createContext<Hedux>(null as any);
+export const HeduxContext = createContext<HeduxContextValue>(null as any);
